fix JaggaerRef.java
