+++ src/main/java/edu/cornell/kfs/module/purap/jaggaer/supplier/xml/JaggaerRef.java
@@ -1,10 +1,39 @@
 package edu.cornell.kfs.module.purap.jaggaer.supplier.xml;
 
-public interface JaggaerRef {
+import jakarta.xml.bind.annotation.XmlElement;
+import jakarta.xml.bind.annotation.XmlTransient;
+
+@XmlTransient
+public abstract class JaggaerRef {
     
-    void setErpNumber(ErpNumber erpNumber);
-    
-    void setSqIntegrationNumber(SQIntegrationNumber sqIntegrationNumber);
-    
-    void setThirdPartyRefNumber(ThirdPartyRefNumber thirdPartyRefNumber);
+    @XmlElement(name = "ERPNumber")
+    private ErpNumber erpNumber;
+    @XmlElement(name = "SQIntegrationNumber")
+    private SQIntegrationNumber sqIntegrationNumber;
+    @XmlElement(name = "ThirdPartyRefNumber")
+    private ThirdPartyRefNumber thirdPartyRefNumber;
+
+    public ErpNumber getErpNumber() {
+        return erpNumber;
+    }
+
+    public void setErpNumber(ErpNumber erpNumber) {
+        this.erpNumber = erpNumber;
+    }
+
+    public SQIntegrationNumber getSqIntegrationNumber() {
+        return sqIntegrationNumber;
+    }
+
+    public void setSqIntegrationNumber(SQIntegrationNumber sqIntegrationNumber) {
+        this.sqIntegrationNumber = sqIntegrationNumber;
+    }
+
+    public ThirdPartyRefNumber getThirdPartyRefNumber() {
+        return thirdPartyRefNumber;
+    }
+
+    public void setThirdPartyRefNumber(ThirdPartyRefNumber thirdPartyRefNumber) {
+        this.thirdPartyRefNumber = thirdPartyRefNumber;
+    }
 }
